Исправленная версия

+++ commerce/delivery/src/main/java/ru/yandex/practicum/delivery/controller/DeliveryController.java
@@ -5,6 +5,8 @@ import ru.yandex.practicum.commerce.interaction.api.dto.DeliveryDto;
 import ru.yandex.practicum.commerce.interaction.api.dto.OrderDto;
 import ru.yandex.practicum.delivery.service.DeliveryService;
 
+import java.math.BigDecimal;
+
 @RestController
 @RequestMapping("/api/v1/delivery")
 public class DeliveryController {
@@ -15,31 +17,41 @@ public class DeliveryController {
         this.deliveryService = deliveryService;
     }
 
-    /** Create a new delivery plan, returning the delivery record with its ID */
+    /**
+     * Create a new delivery plan, returning the delivery record with its ID
+     */
     @PostMapping
     public DeliveryDto planDelivery(@RequestBody OrderDto order) {
         return deliveryService.planDelivery(order);
     }
 
-    /** Calculate delivery cost based on weight, volume, fragility, and addresses */
+    /**
+     * Calculate delivery cost based on weight, volume, fragility, and addresses
+     */
     @PostMapping("/cost")
-    public Double deliveryCost(@RequestBody OrderDto order) {
+    public BigDecimal deliveryCost(@RequestBody OrderDto order) {
         return deliveryService.deliveryCost(order);
     }
 
-    /** Accept order into delivery — sets IN_PROGRESS, notifies order+warehouse */
+    /**
+     * Accept order into delivery — sets IN_PROGRESS, notifies order+warehouse
+     */
     @PostMapping("/pickup")
     public void pickUp(@RequestBody OrderDto order) {
         deliveryService.pickUp(order);
     }
 
-    /** Mark delivery as DELIVERED and notify order service */
+    /**
+     * Mark delivery as DELIVERED and notify order service
+     */
     @PutMapping("/success")
     public void successDelivery(@RequestBody OrderDto order) {
         deliveryService.successDelivery(order);
     }
 
-    /** Mark delivery as FAILED and notify order service */
+    /**
+     * Mark delivery as FAILED and notify order service
+     */
     @PutMapping("/failed")
     public void failedDelivery(@RequestBody OrderDto order) {
         deliveryService.failedDelivery(order);
